services/agents/requirements.txt
fastapi==0.114.0
uvicorn[standard]==0.30.6
GitPython==3.1.43
regex==2024.7.24
requests==2.32.3
